// src/app/languages/interface/more/addAddress.interface.ts
export interface IAddAddress {
    addNewAddress: string;
    selectType: string;   
    home: string;
    office: string;
    other: string;
    country: string;
    state: string;
    fullName: string;
    mobileNumber: string;
    pinCode: string;
    flatNo: string;
    area: string;
    landmark: string;
    town: string;
    submit: string; 
}